pyproject.toml
[project]
name = "obsidian-api-mcp-server"
version = "1.0.1"
description = "MCP server enabling AI agents to perform natural knowledge discovery and analysis across Obsidian vaults"
readme = "README.md"
authors = [
    {name = "pmmvr", email = "17574355+pmmvr@users.noreply.github.com"}
]
license = {text = "MIT"}
keywords = ["obsidian", "mcp", "ai", "knowledge-management", "rest-api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General",
    "Topic :: Office/Business :: Groupware",
]

requires-python = ">=3.10"
dependencies = [
    "mcp>=1.9.1",
    "mcp[cli]",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "behave>=1.2.6",
    "pytest-mock>=3.12.0",
    "responses>=0.24.0",
]

[project.scripts]
obsidian-api-mcp = "obsidian_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_mcp"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.dependencies]
dev = [
    "build>=1.2.2.post1",
    "twine>=6.1.0",
]
